Handle errors when adding dependencies
This PR changes dependency handler to check for existing
artifacts and issue an update if it already exists.

It now also shows an error if a newer version already exists
or dependency cannot be added for whatever reason.

Unfortunately, this change requires going through the gradle
API rather than the generic API as generic API does not provide
an option to update dependency. A nice side effect is that
the gradle api also allows sync